requirements.txt
streamlit
google-generativeai
python-dotenv
PyMuPDF
langchain
langchain-google-genai
chromadb
langchain-community 
langchain-core
langchain-chroma